refactor(calculator): calculator class를 만들어 로직을 분리하라
